requirements.txt
assertpy==1.1
astroid==2.5.6
autopep8==1.5.7
bcrypt==3.1.6
cffi==1.14.5
click==8.0.1
cryptography==3.4.7
dataclasses-json==0.5.4
envyaml==1.8.210417
Flask==1.0.2
Flask-Cors==3.0.9
Flask-JWT==0.3.2
gunicorn==19.9.0
isort==5.9.1
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
marshmallow==3.12.1
marshmallow-enum==1.5.1
mccabe==0.6.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.25
protobuf==3.17.3
pycodestyle==2.7.0
pycparser==2.20
PyJWT==1.4.2
pylint==2.8.3
PyMySQL==1.0.2
python-dotenv==0.10.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.1
stringcase==1.2.0
tabulate==0.8.9
toml==0.10.2
typing-extensions==3.10.0.0
typing-inspect==0.7.1
Werkzeug==2.0.1
wrapt==1.12.1
yoyo-migrations==7.3.2
